+++ src/main/java/com/wds/filmverwaltungsapp/ots/filmverwaltungsappotsrestjavaspringcom/model/FilmInfoModel.java
@@ -14,7 +14,7 @@ public class FilmInfoModel {
 
     private int film_ID;
     private String email;
-    private int bewertung;
+    private double bewertung;
     private boolean favorit;
     private int speichermedien_id;
     private String speichermedium;
